Ajout de composant transforms et physic une meilleure gestion de l'archi et l'enemi tire des lasers

+++ App/controllers/IController.hpp
@@ -58,7 +58,7 @@ public:
 
     virtual ViewTextComponent* newTextRenderer(std::string text);
 
-    virtual Entity* instantiate(Entity* entity, sf::Vector2f position, sf::Vector2f direction){};
+    virtual Entity* instantiate(Entity* entity, sf::Vector2f position, sf::Vector2i direction){};
     virtual void destroy(Entity*entity){};
